// typescript/interfaces/git-hub-stats-dto.ts
export interface GitHubStatDto {
  username: string;
  fullName: string;
  accountAge: string;
  prCount: string;
  commitCount: string;
  forkedCount: string;
  starsCount: string;
  totalRepositories: string;
  totalReposViewsLastTwoWeek: string;
  contributionCount: string;
  totalLinesChanged: string;
}